Adding the crime-baltimore.ipynb file
For the 'Machine Learning for Public Administration' class offered by Korea University's Department of Public Administration. We attempted to solve a societal problem. To reduce fatalities associated to crimes, we decided to create machine learning models that would be able to predict whether a crime would be violent or non-violent. By optimising a DecisionTreeClassifier and RandomForestClassifier, the models reach an accuracy of 0.7976 and an F1-score of ~0.79. Further optimization of models is possible. 

Furthermore, due to insufficient computational power, the KNeighbors model + MLPClassifier model could not be run in the allocated time.